requirements.txt
# StartupGuru - Complete Requirements
# AI & ML Packages
groq==0.9.0
sentence-transformers==2.2.2
numpy==1.26.4
scipy==1.12.0

# Database & Vector Store
chromadb==0.4.15
chromadb-client==0.4.15

# Web Framework & API
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0

# Frontend & UI
streamlit==1.28.0

# Web Scraping & Processing
beautifulsoup4==4.12.2
lxml==4.9.3

# Text Processing
langchain==0.0.350
langchain-text-splitters==0.0.1

# Data Processing
pandas==2.1.3

# Logging & Utils
loguru==0.7.2
python-dotenv==1.0.0
pydantic==2.5.0

# Additional Dependencies
httpx==0.25.2
typing-extensions==4.8.0
jsonschema==4.20.0
pathlib2==2.3.7

# Development & Testing (Optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0
